pyproject.toml
[project]
name = "pose-estimation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.21",
    "rerun-sdk[viewer]>=0.23.1",
    "yt-dlp>=2025.8.27",
]
